% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdData.R
\docType{methods}
\name{as.matrix,ibdData-method}
\alias{as.matrix,ibdData-method}
\title{ibdData as matrix}
\usage{
\S4method{as.matrix}{ibdData}(x, ...)
}
\arguments{
\item{x}{ibdData object}

\item{...}{not implemented}
}
\description{
ibdData as matrix
}

