% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chiSqTest.R
\name{chiSqTest.ibd}
\alias{chiSqTest.ibd}
\title{Chi square test}
\usage{
chiSqTest.ibd(x, spop = rep(1, nrow(x)))
}
\arguments{
\item{x}{ibdData object}

\item{spop}{sub-population index}
}
\description{
Chi square test for IBD sub population data
}

