% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSqTest.R
\name{chiSqTest2Loci.ibd}
\alias{chiSqTest2Loci.ibd}
\title{Chi square test}
\usage{
chiSqTest2Loci.ibd(x, locusI, locusII, spop = rep(1, nrow(x)))
}
\arguments{
\item{x}{ibdData object}

\item{locusI}{chr and position locus I}

\item{locusII}{chr and position locus II}

\item{spop}{sub-population index}
}
\description{
Chi square test for IBD sub population data
}
