% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdData.R
\docType{methods}
\name{dimnames,ibdData-method}
\alias{dimnames,ibdData-method}
\title{Get dimnames}
\usage{
\S4method{dimnames}{ibdData}(x)
}
\arguments{
\item{x}{ibdData object}
}
\description{
Retrieve or set the dimnames of an object.
}
