% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ibdData.R
\docType{methods}
\name{dimnames<-,ibdData,list-method}
\alias{dimnames<-,ibdData,list-method}
\title{Assign dimnames}
\usage{
\S4method{dimnames}{ibdData,list}(x) <- value
}
\arguments{
\item{x}{ibdData object}

\item{value}{List of names}
}
\description{
Retrieve or set the dimnames of an object.
}
\details{


...
}

