% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ibdData-class}
\alias{ibdData-class}
\title{Class "ibdData" for a IBD-block matrix (S4 class)}
\description{
A S4 class for a SNP matrix. Storing SNP information using a 

byte-level (raw) storage scheme jointly with positions.
}
\section{Slots}{

\describe{
\item{\code{ibdData}}{A matrix of IBD-block stored in type 'raw'.}

\item{\code{indivNames}}{A vector of individual names.}

\item{\code{lociNames}}{A vector of loci names.}

\item{\code{alleleLabels}}{A vector of the block-allele label names.}

\item{\code{position}}{A matrix with three rows. The first row contains the 
chromosomes, the second row contains the start positions of a block, and the 
third row contains the end population of a block.}
}}

