% Generated by roxygen2 (4.0.1): do not edit by hand
\name{popstr.matrix}
\alias{popstr.matrix}
\title{design matrix parent population}
\usage{
popstr.matrix(names, sspop, par.list)
}
\arguments{
\item{sspop}{}

\item{par.list}{}

\item{hybrides}{}
}
\description{
design matrix parent population
}

