% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ibdData.R
\name{read.ibdData}
\alias{read.ibdData}
\title{read snp data as ibdData object}
\usage{
read.ibdData(file, header = TRUE, sep = " ",
  quote = "\\"", row.names = FALSE, position.ncol = 3L, na.strings = "NA", nrows = -1)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{header}{a logical value indicating whether the file contains the
names of the variables as its first line}

\item{sep}{the field separator character. Values on each line of the file
are separated by this character.}

\item{quote}{the set of quoting characters. To disable quoting altogether,
use quote = "".}

\item{row.names}{a number giving the column of the table which contains
the row names.}

\item{position.ncol}{integer value indicating whether the how many columns
contain position information.}

\item{na.strings}{a character vector of strings which are to be interpreted
as NA values.}

\item{nrows}{integer: the maximum number of rows to read in.}
}
\description{
read snp data as ibdData object
}

