% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdData.R
\docType{methods}
\name{[,ibdData,ANY,ANY,missing-method}
\alias{[,ibdData,ANY,ANY,missing-method}
\title{Sub ibdData}
\usage{
\S4method{[}{ibdData,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{ibdData object}

\item{i}{indices specifying elements to extract or replace. Indices are n
umeric or character vectors}

\item{j}{indices specifying elements to extract or replace. Indices are 
numeric or character vectors}

\item{...}{not implemented}

\item{drop}{not implemented}
}
\description{
Sub ibdData
}
