% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdData.R
\name{subsetPos}
\alias{subsetPos}
\title{IBD subseting by position}
\usage{
subsetPos(x, chr, pos)
}
\arguments{
\item{x}{ibdData object}

\item{chr}{interger vector for chromosomes}

\item{pos}{interger vector for positions at chr}
}
\description{
IBD subseting by position
}

